requirements.txt
schedule
selenium
python_dotenv
telebot
pyTelegramBotAPI
fake_useragent
undetected_chromedriver